refactor: pages 파일정리 / DP-96 / DP-96
- 인증과 관련된 파일들의 정리실행
- 아이디 찾기, 비밀번호 찾기, 로그인, 회원가입 페이지를 Auth 폴더로 이동
