+++ src/pages/Auth/SignUpPage/SignUpCompletePage.module.scss
