+++ src/router/routes/auth/find-password/change.tsx
@@ -1,6 +1,6 @@
 import { createRoute } from '@tanstack/react-router';
 import { findPasswordRoute } from './index';
-import ChangePasswordPage from '@/pages/FindPasswordPage/ChangePasswordPage';
+import ChangePasswordPage from '@/pages/Auth/FindPasswordPage/ChangePasswordPage';
 
 export const changePasswordRoute = createRoute({
   getParentRoute: () => findPasswordRoute,
