+++ src/router/routes/auth/find-password/index.tsx
@@ -1,6 +1,6 @@
 import { createRoute } from '@tanstack/react-router';
 import { authLayoutRoute } from '../auth-layout';
-import FindPasswordPage from '@/pages/FindPasswordPage/FindPasswordPage';
+import FindPasswordPage from '@/pages/Auth/FindPasswordPage/FindPasswordPage';
 import { changePasswordRoute } from './change';
 
 export const findPasswordRoute = createRoute({
