+++ src/router/routes/auth/sign-in.tsx
@@ -1,6 +1,6 @@
 import { createRoute } from '@tanstack/react-router';
 import { authLayoutRoute } from './auth-layout';
-import SignInPage from '@/pages/SignInPage/SignInPage';
+import SignInPage from '@/pages/Auth/SignInPage/SignInPage';
 
 export const signInRoute = createRoute({
   getParentRoute: () => authLayoutRoute,
