+++ src/router/routes/auth/sign-up/complete.tsx
@@ -1,6 +1,6 @@
 import { createRoute } from '@tanstack/react-router';
 import { signUpRoute } from './index';
-import SignUpCompletePage from '@/pages/SignUpPage/SignUpCompletePage';
+import SignUpCompletePage from '@/pages/Auth/SignUpPage/SignUpCompletePage';
 
 export const signUpCompleteRoute = createRoute({
   getParentRoute: () => signUpRoute,
