+++ src/router/routes/auth/sign-up/index.tsx
@@ -1,6 +1,6 @@
 import { createRoute } from '@tanstack/react-router';
 import { authLayoutRoute } from '../auth-layout';
-import SignUpPage from '@/pages/SignUpPage/SignUpPage';
+import SignUpPage from '@/pages/Auth/SignUpPage/SignUpPage';
 import { signUpCompleteRoute } from './complete';
 
 export const signUpRoute = createRoute({
